requirements.txt
# Nvidia Riva client (ASR + TTS gRPC)
nvidia-riva-client>=2.16.0

# LLM via OpenAI-compatible API
openai>=1.40.0

# GPU monitoring
pynvml>=11.5.0

# Audio processing
soundfile>=0.12.1
numpy>=1.24.0

# Word Error Rate
jiwer>=3.0.4

# Data handling
pandas>=2.0.0
datasets>=2.20.0

# HTTP
httpx>=0.27.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Progress
tqdm>=4.66.0

# UI
streamlit>=1.36.0
plotly>=5.22.0
altair>=5.3.0

# gRPC (pulled in by riva-client but pinned for stability)
grpcio>=1.64.0
protobuf>=4.25.0
